refactor Move

+++ transactions.go
@@ -160,7 +160,7 @@ type PPTransaction struct {
 }
 
 func (t PPTransaction) Mutate(b *Battle) {
-	t.Move.CurrentPP += t.Amount
+	t.Move.CurrentPP += uint8(t.Amount)
 	if t.Move.CurrentPP < 0 {
 		t.Move.CurrentPP = 0
 	} else if t.Move.CurrentPP > t.Move.MaxPP {
